backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.5.2",
    "chainlit>=2.4.1",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "huggingface-hub>=0.29.3",
    "numpy>=2.2.4",
    "ollama>=0.4.7",
    "pathlib>=1.0.1",
    "pydantic>=2.10.6",
    "requests>=2.32.3",
    "sentence-transformers>=3.4.1",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "uvicorn>=0.34.0",
]

[tool.setuptools]
packages = ["backend"]



